# api/openapi.json
{
  "components": {
    "schemas": {
...
      "Availability": {
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "checked_at": {
            "format": "date-time",
            "title": "Checked At",
            "type": "string"
          },
          "reason": {
            "description": "Human-readable: why the provider can or cannot run here",
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "available",
          "reason",
          "checked_at"
        ],
        "title": "Availability",
        "type": "object"
      },
...
      "Citation": {
        "properties": {
          "doi": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doi"
          },
          "pmcid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pmcid"
          },
          "pmid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pmid"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "text",
          "title",
          "year",
          "pmid",
          "pmcid",
          "doi",
          "url"
        ],
        "title": "Citation",
        "type": "object"
      },
...
      "ExecutionMode": {
        "enum": [
          "retrieval",
          "remote_api",
          "local_cli",
          "gpu_worker"
        ],
        "title": "ExecutionMode",
        "type": "string"
      },
...
      "FieldProblem": {
        "properties": {
          "location": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          }
        },
        "required": [
          "location",
          "message"
        ],
        "title": "FieldProblem",
        "type": "object"
      },
...
      "ModelsResponse": {
        "properties": {
          "providers": {
            "items": {
              "$ref": "#/components/schemas/ProviderInfo"
            },
            "title": "Providers",
            "type": "array"
          }
        },
        "required": [
          "providers"
        ],
        "title": "ModelsResponse",
        "type": "object"
      },
...
      "ProblemDetail": {
        "additionalProperties": true,
        "description": "Body of every error response.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldProblem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
      "ProviderInfo": {
        "properties": {
          "attribution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attribution"
          },
          "availability": {
            "$ref": "#/components/schemas/Availability"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "citation": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "title": "Citation",
            "type": "array"
          },
          "commercial_use": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "Whether the license permits commercial use",
            "title": "Commercial Use"
          },
          "execution_mode": {
            "$ref": "#/components/schemas/ExecutionMode"
          },
          "homepage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Homepage"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "job_kinds": {
            "description": "Job kinds that run this provider",
            "items": {
              "type": "string"
            },
            "title": "Job Kinds",
            "type": "array"
          },
          "kind": {
            "$ref": "#/components/schemas/ProviderKind"
          },
          "license": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License"
          },
          "license_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Url"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "max_residues": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Residues"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null when not known on this deployment",
            "title": "Model Version"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "performs_inference": {
            "description": "False for retrieval of an existing prediction",
            "title": "Performs Inference",
            "type": "boolean"
          },
          "requires_api_key": {
            "default": false,
            "title": "Requires Api Key",
            "type": "boolean"
          },
          "requires_gpu": {
            "default": false,
            "title": "Requires Gpu",
            "type": "boolean"
          },
          "structure_origin": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StructureOrigin"
              },
              {
                "type": "null"
              }
            ],
            "description": "Origin carried by structures this provider returns"
          }
        },
        "required": [
          "id",
          "name",
          "kind",
          "model_name",
          "model_version",
          "license",
          "license_url",
          "commercial_use",
          "capabilities",
          "execution_mode",
          "performs_inference",
          "structure_origin",
          "max_residues",
          "requires_api_key",
          "requires_gpu",
          "availability",
          "limitations",
          "citation",
          "attribution",
          "homepage",
          "job_kinds"
        ],
        "title": "ProviderInfo",
        "type": "object"
      },
      "ProviderKind": {
        "enum": [
          "structure_predictor",
          "binding_predictor",
          "variant_effect",
          "pocket",
          "literature"
        ],
        "title": "ProviderKind",
        "type": "string"
      },
...
      "StructureOrigin": {
        "description": "Three structure classes that are never presented as equivalent.",
        "enum": [
          "experimental",
          "predicted_external",
          "predicted_orphafold"
        ],
        "title": "StructureOrigin",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "Open research platform for computational rare-disease work. Research and hypothesis generation\nonly: nothing served here is clinical advice.\n\nEvery error is problem JSON (`application/problem+json`). Aggregated responses carry\n`sources: SourceStatus[]`, so a page renders when one upstream source is down. Writes need the\n`X-OrphaFold-Workspace` header (an anonymous ID generated by the client).\n",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "OrphaFold API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/models": {
      "get": {
        "operationId": "list_models_api_v1_models_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Problem JSON"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Problem JSON"
          }
        },
        "summary": "Model providers, capabilities and availability",
        "tags": [
          "models"
        ]
      }
    },
    "/api/v1/models/{provider_id}": {
      "get": {
        "operationId": "get_model_api_v1_models__provider_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Problem JSON"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Problem JSON"
          }
        },
        "summary": "One model provider",
        "tags": [
          "models"
        ]
      }
    }
  }
}